requirements.txt
google-generativeai
langchain
langchain-community
PyPDF2
faiss-cpu
python-dotenv
tiktoken
